requirements.txt
langchain==1.1.0
langchain-community==0.4.1
langchain-core==1.1.3
langchain-google-genai==1.0.8
langchain-anthropic==1.2.0
python-dotenv==1.2.1
pydantic==2.12.5
duckduckgo-search==8.1.1
wikipedia==1.4.0
fastapi==0.128.0
uvicorn==0.40.0
